checkpoint : a customer can place an order

+++ OrderDomain/src/main/java/za/co/applications/princegains/shopping/shopping/service/OrderService.java
@@ -19,5 +19,5 @@ public interface OrderService {
 
     void updateAnOrder(Order order);
 
-    Order getAnOrder(long id);
+    Order getAnOrder(int id);
 }
